% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Financial_Statement_Analysis.R
\name{getBalanceSheet}
\alias{getBalanceSheet}
\title{Return data from balance sheet}
\usage{
getBalanceSheet(company, periodFreq)
}
\arguments{
\item{company}{data from getFinancials functions;}

\item{period}{can be 'Q' for quarterly data and 'A' for annual data.}
}
\value{
Returns dataFrame type with balance sheet data
}
\description{
Return data from balance sheet
}

