% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Financial_Statement_Analysis.R
\name{getCashFlow}
\alias{getCashFlow}
\title{Return data from Cash Flow Statement}
\usage{
getCashFlow(company, periodFreq)
}
\arguments{
\item{company}{data from getFinancials funcitons; period can be 'Q' for quarterly data and 'A' for annual data.}
}
\value{
Returns dataFrame type with cash flow data
}
\description{
Return data from Cash Flow Statement
}

