% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Financial_Statement_Analysis.R
\name{getIncomeStatement}
\alias{getIncomeStatement}
\title{Return data from Income statement}
\usage{
getIncomeStatement(company, periodFreq)
}
\arguments{
\item{company}{data from getFinancials functions; period can be 'Q' for quarterly data and 'A' for annual data.}
}
\value{
Returns dataFrame type with income statement data
}
\description{
Return data from Income statement
}

